pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reportobello"
version = "0.2.1"
description = "Python API library for Reportobello"
authors = [{name = "dosisod"}]
license = {text = "LGPL-3.0-or-later"}
readme = "README.md"
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
  "Typing :: Typed"
]
requires-python = ">=3.9"
dependencies = [
  "httpx >= 0.27.2",
  "rich >= 13.8.0",
  "typst >= 0.11.1",
  "python-dotenv >= 1.0.1",
]

[project.scripts]
rpbl = "cli:main"

[project.urls]
Source = "https://github.com/reportobello/reportobello-python"

[tool.setuptools]
py-modules = ["cli", "reportobello"]
